Moved closures from the main runtime file
