+++ runtime/src/runtime.js
@@ -17,41 +17,6 @@ export const encoder = new TextEncoder("utf-8");
 export const decoder = new TextDecoder("utf-8");
 let counter = 0;
 
-var GrainClosure = function(loc) {
-  this.loc = loc;
-  this.arity = view[loc];
-  this.ptr = view[loc + 1];
-  this.closureSize = view[loc + 2];
-  this.closureElts = view.slice(loc + 3, loc + 3 + this.closureSize);
-  this.func = grainModule.instance.exports["GRAIN$LAM_" + this.ptr];
-};
-
-GrainClosure.prototype.call = function() {
-  if (arguments.length != this.arity) {
-    throwGrainError(GRAIN_ERR_ARITY_MISMATCH, this.arity, arguments.length);
-    return undefined;
-  } else {
-    let grainVals = Array.prototype.map.call(arguments, JSToGrainVal);
-    grainVals.unshift(this.loc * 4);
-    return grainToJSVal(this.func.apply(this.func, grainVals));
-  }
-};
-
-function printClosure(c) {
-  c /= 4;
-  let arity = view[c];
-  let idx = view[c + 1];
-  let closureSize = view[c + 2];
-  let closureElts = [];
-
-  for (var i = 0; i < closureSize; ++i) {
-    closureElts.push(printNumber(view[c + i + 3]));
-  }
-  console.log(`<closure@${c}: idx=${idx}, arity=${arity}, size=${closureSize}: ${closureElts}>`);
-  console.log(view.slice(0, 32));
-  return c;
-}
-
 function grainEqualHelp(x, y, cycles) {
   if ((x & 7) === 1) {
     if ((y & 7) !== 1) {
